PIMixin: removed support for using a timeseries_export.xml placeholder file. This file is always entirely constructed by RTC-Tools 2. This also means that we have removed support for interpolation of the output series (this was based on the times in the output placeholder). This should be done in FEWS.
The MixIn now outputs all variables in the output_variables list to the
timeseries_export, if the variable has a location and parameter mapping in
rtcDataConfig. If not, the variable is skipped.

Furthermore, the mixin of course passes all relevant information from the
import timseries and rtcDataConfig to the export file (times, start/end
date, location, parameter, etc.) to be able to construct a fully filled
TimeSeries object.

Added that the timezone of the import file is read and transferred to the
export file.

Added proper support for writing ensemble values (previously this did not
function as intended).

Added that the forecast datetime is writting in the export file if it was
present in the import file.

Added mapping of RTCTools variable to FEWS location/parameter/qualifiers (the reverse
was already present).