Add GitLab CI file for wheels and tests
So far only (binary) wheels on Windows for Python 2.7 32-bits are
supported. No builds are performed for Linux (or MacOS) yet, or for other
versions of Python.

For tests a Docker environment on Linux is used, as this will allow using
that same runner to easily build linux wheels in the future (see PEP-513).

NOTE: For some reason, using a GitLab runner with version 1.8.0 does not
work, whereas 1.7.1 does. The former cannot find local images, which is
possibly related to some default pull policy.