Disallow multiple goals with the same key and priority
If two goals have the same priority and function key, the updating of
min/max contraints for this state/function_key throughout the priorities
will not work correctly. Instead of working around such cases, we want
users to aggregate these (possibly conflicting) goals before optimizing.
Therefore, an exception will be raised notifying the user of the two
goals that should be merged.