Fix initial state reading from aliases in CSVMixin
Also update the test case to set the initial state with an alias. This
new test case would have failed without the CSVMixin changes made in this
commit.

Note that the test case for PIMixin has also been updated to keep matching
the other test case, but that the implementation of aliases in PIMixin
was already correct.

Regression was introduced in 989632b84d6f27cabb117f5222ad3a18b5bd8503.