Remove broken history() in ModelicaMixin
If history() is override, a user must take care to also explicitly call
self.initial_state(). In the history() method that was removed, this was
_not_ the case, resulting in missing initial state values specified in
e.g. the initial_state.csv file.

Furthermore, the behavior of the history() method matched the default
behavior of the inherited history() (from OptimizationProblem) and
the already present initial_state().