Temporarily disable alias detection in compiler
Alias detection on the compiler level strips data that we need. E.g. when
we specify a desired output "V_storage", which is an alias of "Storage.V",
we could end up with an output of "Storage.V" instead. We would want to
use the alias the user wants to output, before enabling alias
detection/reduction again.