Rename distribution name to rtc-tools
The import name of the package remains "rtctools" as no dashes are
supported in those.

Note that the "rtctools" package on PyPI has been made an alias to the
"rtc-tools" package. That way we show the preference for the usage of
the dash version, but still allow users to make a typo in trying to
install "rtctools" instead.