Relicense to LGPLv3, and remove GPL notices
The filenames COPYING and COPYING.LESSER are suggested by the Free
Software Foundation, and also typically used in many automatic checkers.

The command-line notices were required by the GPL, but are not needed to
be compliant with the LGPL. We remove them to behave more like similarly
licensed software (e.g. CasADi), which do not produce such notices.