Fix reading ensemble csv and bump NumPy version
Starting with NumPy 1.14, the encoding of reading the file can be
specified. The default 'bytes' enables backwards compatibility, but for
some reason explicit conversion to string using the converters does not
work anymore. Instead of working around the issue and support both older
and newer version of NumPy, we bump the version requirement and use the
new "encoding" argument.