Do not use variables for min/max of path goals
Using variables for the min/max series of path goals resulted in (at
least) quadratic constraints. These min/max series variables had equal
bounds, and most solvers remove these variables in the preprocessor of the
optimization. So although there was no problem for e.g. IPOPT, it was not
possible for the user or CasADi to easily detect whether the problem was
linear.

As a concrete example, it was not possible to use CLP (a linear solver),
because a Hessian check in CasADi would fail. Furthermore, there might be
solvers that we want to use that do not (by default) support the
elimination of variables with equal bounds.

Instead of using variables, we can instead specify the min/max series
as additional constant inputs.

Fixes #1005