Update casadi3 branch references to master
The casadi3 branch was the point of main development, and has now
replaced master. The latter has been renamed/moved to maintenance/2.0.x.
Any references to the casadi3 branch in the code (code coverage, badges)
have now been update to point to master.

Note that the actual renaming of casadi3 to master was a "simple" force
push, and not a merge commit. This is because development had diverged
for too long, and no active development was happening on the "old"
master anyway.