Change flake8 max line length to 120
Much of RTC-Tools's code is >100 already, with many long variable names
being used. Aiming for 99 chars will therefore only tend to make code less
readable. The 120 chararacter limit is the maximum line length for GitHub
and Gitlab, and the default in e.g. PyCharm, so it seems like the obvious
choice.