Add support for expansion of nlp
In CasADi 2 the solver options were at the same level as the CasADi
options. In CasADi 3, the options to be passed to e.g. IPOPT need to be
one level deeper.

As we only need very few options to CasADi's nlpsol/qpsol, we make a
special case out of them. This was already the case for e.g. "discrete"
and "iteration_callback". However, "expand" fell by the wayside in the
conversion from CasADi 2 to 3. It was however actively used in RTC-Tools
2.0 configurations.

Instead of simply setting "expand" to True at the level of CasADi options,
we do the work ourselves. This is because only CasADi's nlpsol support the
expand option, and e.g. qpsol does not. Note that the user cannot yet
choose between nlpsol and qpsol, but that this feature is planned.